src/requirements.txt
rasa==3.6.21
fastapi==0.118.3
uvicorn==0.37.0
python-dotenv==1.1.1